Big Update screen.html
wow I added a lot... I changed the input boxes and buttons to now be a slightly darker theme in light mode, and lighter in dark mode. I also changed the text from yellow to black or white depending on light or dark mode. I also added an exclamation mark at the end of pizzazz, added a favicon, fixed some bugs, and changed the logo! a lot was done so I hope you enjoy